requirements.txt
fastapi
pydantic
requests
uvicorn
supabase
asyncpraw
httpx
python-dotenv
psutil